Move replication resources to own package
Try to fix 503 errors with retry
