+++ pkg/artifactory/resource/replication/replication.go
@@ -1,0 +1,3 @@
+package replication
+
+const ReplicationEndpointPath = "artifactory/api/replications/"
