+++ pkg/artifactory/resource/replication/resource_artifactory_single_replication_config_test.go
@@ -1,28 +1,30 @@
-package artifactory_test
+package replication_test
 
 import (
 	"fmt"
+	"log"
 	"os"
 	"regexp"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/acctest"
 )
 
-func mkTclForPullRepConfg(name, cron, url string) string {
+func mkTclForRepConfg(name, cron, url, proxy string) string {
 	const tcl = `
 		resource "artifactory_local_maven_repository" "%s" {
 			key = "%s"
 		}
 
-		resource "artifactory_pull_replication" "%s" {
+		resource "artifactory_single_replication_config" "%s" {
 			repo_key = "${artifactory_local_maven_repository.%s.key}"
 			cron_exp = "%s"
 			enable_event_replication = true
 			url = "%s"
 			username = "%s"
-			password = "Passw0rd!"
+			proxy = "%s"
 		}
 	`
 	return fmt.Sprintf(tcl,
@@ -33,13 +35,13 @@ func mkTclForPullRepConfg(name, cron, url string) string {
 		cron,
 		url,
 		acctest.RtDefaultUser,
+		proxy,
 	)
 }
+func TestInvalidCronSingleReplication(t *testing.T) {
 
-func TestAccPullReplicationInvalidCron(t *testing.T) {
-
-	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_pull_replication")
-	var failCron = mkTclForPullRepConfg(name, "0 0 * * * !!", os.Getenv("ARTIFACTORY_URL"))
+	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_single_replication_config")
+	var failCron = mkTclForRepConfg(name, "0 0 * * * !!", os.Getenv("ARTIFACTORY_URL"), "")
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { acctest.PreCheck(t) },
@@ -54,10 +56,66 @@ func TestAccPullReplicationInvalidCron(t *testing.T) {
 	})
 }
 
-func TestAccPullReplicationLocalRepo(t *testing.T) {
-	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_pull_replication")
-	config := mkTclForPullRepConfg(name, "0 0 * * * ?", os.Getenv("ARTIFACTORY_URL"))
-	updatedConfig := mkTclForPullRepConfg(name, "1 0 * * * ?", os.Getenv("ARTIFACTORY_URL"))
+func TestInvalidUrlSingleReplication(t *testing.T) {
+
+	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_single_replication_config")
+	var failCron = mkTclForRepConfg(name, "0 0 * * * ?", "bad_url", "")
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { acctest.PreCheck(t) },
+		ProviderFactories: acctest.ProviderFactories,
+		CheckDestroy:      testAccCheckReplicationDestroy(fqrn),
+		Steps: []resource.TestStep{
+			{
+				Config:      failCron,
+				ExpectError: regexp.MustCompile(`.*expected "url" to have a host, got bad_url.*`),
+			},
+		},
+	})
+}
+
+func TestAccSingleReplication_full(t *testing.T) {
+	const testProxy = "testProxy"
+	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_single_replication_config")
+	config := mkTclForRepConfg(name, "0 0 * * * ?", os.Getenv("ARTIFACTORY_URL"), testProxy)
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			acctest.PreCheck(t)
+			acctest.CreateProxy(t, testProxy)
+		},
+		ProviderFactories: acctest.ProviderFactories,
+		CheckDestroy: func() func(*terraform.State) error {
+			acctest.DeleteProxy(t, testProxy)
+			return testAccCheckReplicationDestroy(fqrn)
+		}(),
+
+		Steps: []resource.TestStep{
+			{
+				Config: config,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "repo_key", name),
+					resource.TestCheckResourceAttr(fqrn, "cron_exp", "0 0 * * * ?"),
+					resource.TestCheckResourceAttr(fqrn, "enable_event_replication", "true"),
+					resource.TestCheckResourceAttr(fqrn, "url", os.Getenv("ARTIFACTORY_URL")),
+					resource.TestCheckResourceAttr(fqrn, "username", acctest.RtDefaultUser),
+					resource.TestCheckResourceAttr(fqrn, "proxy", testProxy),
+				),
+			},
+		},
+	})
+}
+
+func TestAccSingleReplication_withDelRepo(t *testing.T) {
+	_, fqrn, name := acctest.MkNames("lib-local", "artifactory_single_replication_config")
+	config := mkTclForRepConfg(name, "0 0 * * * ?", os.Getenv("ARTIFACTORY_URL"), "")
+	var deleteRepo = func() {
+		restyClient := acctest.GetTestResty(t)
+		_, err := restyClient.R().Delete("artifactory/api/repositories/" + name)
+		if err != nil {
+			t.Fatal(err)
+		}
+		log.Printf("Delete repo %s done.", name)
+	}
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { acctest.PreCheck(t) },
 		ProviderFactories: acctest.ProviderFactories,
@@ -70,26 +128,24 @@ func TestAccPullReplicationLocalRepo(t *testing.T) {
 					resource.TestCheckResourceAttr(fqrn, "repo_key", name),
 					resource.TestCheckResourceAttr(fqrn, "cron_exp", "0 0 * * * ?"),
 					resource.TestCheckResourceAttr(fqrn, "enable_event_replication", "true"),
+					resource.TestCheckResourceAttr(fqrn, "url", os.Getenv("ARTIFACTORY_URL")),
 					resource.TestCheckResourceAttr(fqrn, "username", acctest.RtDefaultUser),
-					resource.TestCheckResourceAttr(fqrn, "password", "Passw0rd!"),
+					resource.TestCheckResourceAttr(fqrn, "proxy", ""),
 				),
 			},
 			{
-				Config: updatedConfig,
+				PreConfig: deleteRepo,
+				Config:    config,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(fqrn, "repo_key", name),
-					resource.TestCheckResourceAttr(fqrn, "cron_exp", "1 0 * * * ?"),
-					resource.TestCheckResourceAttr(fqrn, "enable_event_replication", "true"),
-					resource.TestCheckResourceAttr(fqrn, "username", acctest.RtDefaultUser),
-					resource.TestCheckResourceAttr(fqrn, "password", "Passw0rd!"),
 				),
 			},
 		},
 	})
 }
 
-func TestAccPullReplicationRemoteRepo(t *testing.T) {
-	_, fqrn, name := acctest.MkNames("lib-remote", "artifactory_pull_replication")
+func TestAccSingleReplicationRemoteRepo(t *testing.T) {
+	_, fqrn, name := acctest.MkNames("lib-remote", "artifactory_single_replication_config")
 	_, fqrepoName, repo_name := acctest.MkNames("lib-remote", "artifactory_remote_maven_repository")
 	var tcl = `
 		resource "artifactory_remote_maven_repository" "{{ .remote_name }}" {
@@ -98,16 +154,19 @@ func TestAccPullReplicationRemoteRepo(t *testing.T) {
 			repo_layout_ref       = "maven-2-default"
 		}
 
-		resource "artifactory_pull_replication" "{{ .repoconfig_name }}" {
+		resource "artifactory_single_replication_config" "{{ .repoconfig_name }}" {
 			repo_key = "{{ .remote_name }}"
 			cron_exp = "0 0 12 ? * MON *"
 			enable_event_replication = false
+			url = "https://repo1.maven.org/maven2/"
+			username = "{{ .username }}"
 			depends_on = [artifactory_remote_maven_repository.{{ .remote_name }}]
 		}
 	`
 	tcl = acctest.ExecuteTemplate("foo", tcl, map[string]string{
 		"repoconfig_name": name,
 		"remote_name":     repo_name,
+		"username":        acctest.RtDefaultUser,
 	})
 
 	resource.Test(t, resource.TestCase{
